Game-Client: run game-notes migration when loading game notes
Performance improvement. Moves "xml" to "html" game-note
migration from program launch to instead be done (on-demand)
when we read game-notes.

This avoids a possible performance hit when launching the game.
'PerfTimer' has the game notes migration process as needing
about 30ms (so we are adding a pretty minimal one-time time
cost to reading game notes)

+++ game-app/map-data/src/main/java/org/triplea/map/game/notes/GameNotesMigrator.java
@@ -1,13 +1,7 @@
 package org.triplea.map.game.notes;
 
-import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Optional;
-import java.util.function.Consumer;
-import java.util.function.Predicate;
-import java.util.stream.Collectors;
 import lombok.Builder;
 import lombok.extern.slf4j.Slf4j;
 import org.triplea.generic.xml.reader.XmlMapper;
@@ -24,54 +18,18 @@ import org.triplea.map.data.elements.PropertyList;
 @Builder
 @Slf4j
 public class GameNotesMigrator {
-
-  /** Callback to be invoked if we start generating YAML files. */
-  private final Consumer<Runnable> progressIndicator;
-
-  /** Path to where downloaded maps can be found. */
-  private final Path downloadedMapsFolder;
-
-  public void extractGameNotes() {
-    final Collection<Path> xmlFilesMissingGameNotesFile =
-        findAllGameXmlFiles(downloadedMapsFolder).stream()
-            .filter(Predicate.not(GameNotes::gameNotesFileExistsForGameXmlFile))
-            .collect(Collectors.toSet());
-
-    // For each game XML missing a game-notes file, generate the game-notes file.
-    // If we have a number of game notes files to create, run in a progress dialog.
-
-    final Runnable migrationJob =
-        () ->
-            xmlFilesMissingGameNotesFile.forEach(
-                xmlGameFile -> {
-                  final String gameNotes = readGameNotesFromXml(xmlGameFile);
-                  final String fileNameToWrite = GameNotes.createExpectedNotesFileName(xmlGameFile);
-                  final Path fileToWrite = xmlGameFile.resolveSibling(fileNameToWrite);
-                  log.info("Writing game notes file: " + fileToWrite.toAbsolutePath());
-                  FileUtils.writeToFile(fileToWrite, gameNotes);
-                });
-
-    if (xmlFilesMissingGameNotesFile.size() > 3) {
-      progressIndicator.accept(migrationJob);
-    } else {
-      migrationJob.run();
+  public static void extractGameNotes(Path xmlGameFile) {
+    if (GameNotes.gameNotesFileExistsForGameXmlFile(xmlGameFile)) {
+      return;
     }
+    final String gameNotes = readGameNotesFromXml(xmlGameFile);
+    final String fileNameToWrite = GameNotes.createExpectedNotesFileName(xmlGameFile);
+    final Path fileToWrite = xmlGameFile.resolveSibling(fileNameToWrite);
+    log.info("Writing game notes file: " + fileToWrite.toAbsolutePath());
+    FileUtils.writeToFile(fileToWrite, gameNotes);
   }
 
-  private Collection<Path> findAllGameXmlFiles(final Path rootPath) {
-    final Collection<Path> xmlFiles = new ArrayList<>();
-
-    for (final Path child : FileUtils.listFiles(rootPath)) {
-      if (Files.isDirectory(child) && child.getFileName().toString().equalsIgnoreCase("games")) {
-        xmlFiles.addAll(FileUtils.findXmlFiles(child, 1));
-      } else if (Files.isDirectory(child)) {
-        xmlFiles.addAll(findAllGameXmlFiles(child));
-      }
-    }
-    return xmlFiles;
-  }
-
-  private String readGameNotesFromXml(final Path xmlGameFile) {
+  private static String readGameNotesFromXml(final Path xmlGameFile) {
     final Game game =
         FileUtils.openInputStream(
             xmlGameFile,
@@ -104,7 +62,7 @@ public class GameNotesMigrator {
         .orElse("");
   }
 
-  private Optional<PropertyList.Property> getGameNotesProperty(final Game game) {
+  private static Optional<PropertyList.Property> getGameNotesProperty(final Game game) {
     return game.getPropertyList().getProperties().stream()
         .filter(property -> property.getName() != null)
         .filter(property -> property.getName().equalsIgnoreCase("notes"))
